Make local API more reliable during development
While working on API changes, the watcher occasionally crashes with:
```
Emitted 'error' event on Server instance at:
    at emitErrorNT (node:net:1899:8)
    at processTicksAndRejections (node:internal/process/task_queues:82:21) {
  code: 'EADDRINUSE',
  errno: -4091,
  syscall: 'listen',
  address: '::',
  port: 6540
}

Node.js v20.11.0
[nodemon] app crashed - waiting for file changes before starting...
```
And gets stuck. It needs to be manually killed along with the port, then the command needs to be run again to get the local API to work.
This commit means that the API-watcher is slightly slower each time but much more reliable.